+++ .gitignore
@@ -26,10 +26,10 @@ apache-tomcat-*
 hieroweb/classes
 
 # Generated by webpack
-hieroweb/web/bundle.js
-hieroweb/web/bundle.js.map
-hieroweb/web/node_modules/
-hieroweb/web/typings/*
+hieroweb/bundle.js
+hieroweb/bundle.js.map
+hieroweb/node_modules/
+hieroweb/typings/*
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
