+++ hieroweb/InitialObject.ts
