+++ hieroweb/src/main/java/org/hiero/RpcReply.java
@@ -1,4 +1,4 @@
-package hiero.web;
+package org.hiero;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
