+++ hieroweb/tsconfig.json
@@ -6,5 +6,5 @@
     "sourceMap": true,
     "target": "es5"
   },
-  "exclude": [ "node_modules" ]
+  "exclude": [ "node_modules", "target", "classes" ]
 }
